move all league stat methods to LeagueStats module
